DDA-197: Move copyFolderRecursiveSync to utils to allow easier mocking
